requirements.txt
requests==2.32.3
pandas==2.2.3
psycopg[binary]==3.2.3
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
tenacity==9.0.0
pytest==8.3.3
pytest-cov==5.0.0
ruff==0.7.4
testcontainers[postgres]==4.8.2
